Replaced min and max macros with template functions

+++ Steam/steam/strtools.cpp
@@ -1994,7 +1994,7 @@ bool V_ExtractFilePath (const char *path, char *dest, int destSize )
 		src--;
 	}
 
-	int copysize = min( src - path, destSize - 1 );
+	int copysize = min( (int)(src - path), destSize - 1 );
 	memcpy( dest, path, copysize );
 	dest[copysize] = 0;
